% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{coef.inlamemi}
\alias{coef.inlamemi}
\title{Extract model coefficients}
\usage{
\method{coef}{inlamemi}(object, ...)
}
\arguments{
\item{object}{object of class `inlamemi`}

\item{...}{other arguments}
}
\description{
Extract model coefficients
}
