% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_random_effects}
\alias{extract_random_effects}
\title{Extract random effects from formula}
\usage{
extract_random_effects(formula)
}
\arguments{
\item{formula}{an object of class "formula", either the formula for the model of interest, the imputation model or the missingness model.}
}
\value{
A list containing "reff_vars", the random effect variables, and "reff", the entire random effect term.
}
\description{
Extract random effects from formula
}
