% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_coef_imp}
\alias{get_coef_imp}
\title{Extract coefficients for the imputation model (IMP)}
\usage{
get_coef_imp(inlamemi_model)
}
\arguments{
\item{inlamemi_model}{object of class `inlamemi`}
}
\value{
A data frame with a summary of the posterior marginals for the coefficients in the imputation model.
}
\description{
Extract coefficients for the imputation model (IMP)
}
