% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_coef_mis}
\alias{get_coef_mis}
\title{Extract coefficients for the missingness model (MIS)}
\usage{
get_coef_mis(inlamemi_model)
}
\arguments{
\item{inlamemi_model}{object of class `inlamemi`}
}
\value{
A data frame with a summary of the posterior marginals for the coefficients in the missingness model.
}
\description{
Extract coefficients for the missingness model (MIS)
}
