% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_coef_moi}
\alias{get_coef_moi}
\title{Extract coefficients for the model of interest (MOI)}
\usage{
get_coef_moi(inlamemi_model)
}
\arguments{
\item{inlamemi_model}{object of class `inlamemi`}
}
\value{
A data frame with a summary of the posterior marginals for the coefficients in the model of interest.
}
\description{
Extract coefficients for the model of interest (MOI)
}
