% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_imputed}
\alias{get_imputed}
\title{Extract imputed values}
\usage{
get_imputed(inlamemi_model, error_variable)
}
\arguments{
\item{inlamemi_model}{object of class `inlamemi`}

\item{error_variable}{character string indicating the name of the error variable for which to extract the imputed values.}
}
\value{
A list of two objects: the posterior marginal distributions for each element of the imputed covariate, and a data frame giving a summary of these marginals.
}
\description{
Extract imputed values
}
