% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mar_data}
\alias{mar_data}
\title{Simulated data with observation missing at random (MAR)}
\format{
## `mar_data`
A data frame with 1000 rows and 5 columns:
\describe{
  \item{y}{Response variable}
  \item{x}{Observed value of covariate, with almost 20 percent missing}
  \item{x_true}{Correct version of x, without missingness}
  \item{z1}{Covariate correlated with x}
  \item{z2}{Covariate correlated with the missingness of x}
}
}
\source{
The dataset is simulated.
}
\usage{
mar_data
}
\description{
A simulated dataset to demonstrate how to set up a model in the case where there are two variables with measurement error.
}
\keyword{datasets}
