% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{print.inlamemi}
\alias{print.inlamemi}
\title{Print method for inlamemi}
\usage{
\method{print}{inlamemi}(x, ...)
}
\arguments{
\item{x}{object of class `inlamemi`.}

\item{...}{other arguments.}
}
\description{
Print method for inlamemi
}
